+++ crates/fluvio-version-manager/src/common/mod.rs
@@ -2,6 +2,7 @@ pub mod manifest;
 pub mod notify;
 pub mod settings;
 pub mod version_directory;
+pub mod version_installer;
 pub mod workdir;
 
 use std::path::PathBuf;
